fixed the query doubled file
